### 2021-05-06 #3258 @whedon
[PRE REVIEW]: dataaimsr: An R Client for the Australian Institute of Marine Science Data Platform API which provides easy access to AIMS Data Platform
**Submitting author:** @dbarneche (<a href="http://orcid.org/0000-0002-4568-2362">Diego Barneche</a>)
**Repository:** <a href="https://github.com/ropensci/dataaimsr" target ="_blank">https://github.com/ropensci/dataaimsr</a>
**Version:** v1.0.3
