### 2021-05-07 #3258 @dbarneche
Hi @kthyng ! Thank you very much for acting as Editor.
Also thanks for flagging the structural issue. We currently have a section named [`Background and Statement of need`](https://github.com/ropensci/dataaimsr/blob/master/paper/paper.md#background-and-statement-of-need), should I rename it to `Statement of need` only?
Have a good weekend.
