### 2021-05-13 #3258 @kthyng
@dbarneche Yes, please rename it. I see your paper is pretty long — this is almost to the length where I would require it to be shorter but we try not to be too heavy handed about it. If some of it is repeat from the docs, you could cut it from the paper. But, if you feel like it is how you'd like it to be, you can leave it.
