### 2021-05-24 #3282 @dbarneche
Hi @kthyng,

Thanks very much for advancing this, and apologies for the delayed response. I have addressed the comments mention in the [pre-review issue](https://github.com/openjournals/joss-reviews/issues/3258). Particularly, I renamed the first section to "Statement of Need" and shortened some bits of the paper as advised. There is still a bit of repetition with the package documentation/vignettes, but I thought those were best as is for the sake of transparency---let me know if you'd like me to shrink the size further.

I have also created a Zenodo repo for the package, please see it here:
https://zenodo.org/record/4782786#.YKsgIpMzbUI

DOI is: 10.5281/zenodo.4782786.svg

Please do not hesitate to let me know if there's anything else needed from my end. Thanks again. Best wishes
