### 2021-05-24 #3282 @arfon
> @openjournals/dev I think I lost my permission to edit the checkboxes above midway through checking them. They were turning blue for me and I could edit and now they are gray and I cannot edit them.

Weird... Can you try again now please?
