### 2021-05-26 #3282 @kthyng
@dbarneche Great! Can you update the title of your Zenodo archive to exactly match your JOSS paper?

Also, the figure references in your JOSS paper seem to have gotten off. Can you take a look at them in the proof above? You don't need to update the Zenodo archive for paper changes (unless you want to).
