### 2021-05-27 #3282 @dbarneche
Hi @kthyng , apologies for the mix up with the Zenodo title, that has been fixed.

I'm not quite sure how to fix the figure references/captions issue. Would you mind guiding me through a fix?

Locally, I'm running [this code](https://github.com/ropensci/dataaimsr/blob/master/paper/render.R) to generate the pdf (it is a bit of a hack), and the fig references/captions render fine, e.g.:

![Screen Shot 2021-05-27 at 9 35 12 am](https://user-images.githubusercontent.com/3580907/119752035-d7eb6200-bece-11eb-92bc-e4788fbefc66.png)
![Screen Shot 2021-05-27 at 9 35 23 am](https://user-images.githubusercontent.com/3580907/119752051-dde14300-bece-11eb-91f7-847d0f4949e7.png)
![Screen Shot 2021-05-27 at 9 35 57 am](https://user-images.githubusercontent.com/3580907/119752104-f2254000-bece-11eb-9b18-6dc71ed2383c.png)
![Screen Shot 2021-05-27 at 9 36 08 am](https://user-images.githubusercontent.com/3580907/119752119-f94c4e00-bece-11eb-85c3-1f17aa711298.png)

Thanks for your help!
