### 2021-06-01 #3282 @kthyng
@dbarneche There is a note about paper compilation here: https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles We have a system that will do that for you, and you can test it out, too. Examples of figures and captions are here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography Does this help?
