### 2021-06-01 #3282 @dbarneche
Thanks @kthyng ! It has been [fixed](https://github.com/ropensci/dataaimsr/commit/b7daff1254a6f7741b7117e2e2115d331494b37a).
I checked the online compilation tools and it should be working now.
